Fixed second column, wagon positioning
I gave up on putting the wagon/emote in the top right. There's something
weird about the CSS that I cannot figure out causing absolute
positioning to fail in the top right corner, but not the top left.
So now it's in the left unless your screen is large enough. It seems to
work really well on any size screen, so after I get feedback I'll merge
this into master.